third_party/blink/web_tests/virtual/disable-ua-client-hint/external/wpt/client-hints/accept-ch-stickiness/same-origin-navigation-redirect.https-expected.txt

This is a testharness.js-based test.
[FAIL] redirect on navigation got client hints according to expectations.
  assert_equals: message from opened page expected "PASS" but got "UA"
Harness: the test ran to completion.
